Use sqrtf for weight scaling in MLP initialization for improved precision

+++ gpu/mlp.c
@@ -27,9 +27,9 @@ MLP* init_mlp(int input_dim, int hidden_dim, int output_dim, int batch_size) {
     float* W3 = (float*)malloc(output_dim * input_dim * sizeof(float));
     
     // Initialize weights on host
-    float scale_W1 = 1.0f / sqrt(input_dim);
-    float scale_W2 = 1.0f / sqrt(hidden_dim);
-    float scale_W3 = 1.0f / sqrt(input_dim);
+    float scale_W1 = 1.0f / sqrtf(input_dim);
+    float scale_W2 = 1.0f / sqrtf(hidden_dim);
+    float scale_W3 = 1.0f / sqrtf(input_dim);
     
     for (int i = 0; i < hidden_dim * input_dim; i++) {
         W1[i] = ((float)rand() / (float)RAND_MAX * 2.0f - 1.0f) * scale_W1;
